+++ Release Notes.md
@@ -27,6 +27,7 @@ SOFTWARE.
 * FM24C & MB85RC256V: added macros to create simple inlines to read/write values from addresses (to use in some header)
 * MTCH6102: added standalone function for commands & inlines for configuration, restore to defaults & tests commands
 * MTCH6102: added filtering type configuration function & inlines for configuration
+* MTCH6102: added average compensation calculation for channels following datasheet & manufacturing test execution with results
 
 ## v0.6